numbered error messages
Error messages are now prefixed with a number, eg "err03". This way the message text itself can change and still an easy check can be made in tests that a certain error got thrown.

Numbering started from the first file in the project. There is no special meaning in the error number (eg importance).